pyproject.toml
[project]
name = "promptyard"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django-allauth[socialaccount]>=65.2.0",
    "django-simple-history>=3.7.0",
    "django-taggit>=6.1.0",
    "django>=5.1.3",
    "pre-commit>=4.0.1",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.0.1",
]

[dependency-groups]
dev = [
    "commitizen>=3.30.1",
    "ruff==0.7.3",
]

[tool.commitizen]
name = "cz_conventional_commits"
tag_format = "$version"
version_scheme = "pep440"
version_provider = "pep621"
update_changelog_on_bump = true
major_version_zero = true
